Make sure all prereqs are declared in code
I removed the explicit lists from dzil - we should bump versions in code when
we need to enforce a particular dep's version. I think this is a better
practice since it means that if someone installs this distro by hand they
will still get errors about not having the right versions of prereqs.